UI_Changes_and_Undead_Rig
Added an undead rig and hooked it up to the ESRPG undead blueprint. Added the enemy to the elven realm gates level for testing. Also tinkered a bit with the UI and added a bunch of controller input support, but still have a ton of controller input left to do.