BrokenButNotBrokenEnough
Need to rework the C# directory files for C++ to compile, so this is not a working version, but there's too much content to risk losing, so I'm committing it anyway. The workaround is to modify the C# pointer files and remove the plugin dependencies until the unknown character errors can be resolved.